+++ convoy/settings.py
@@ -187,7 +187,8 @@ PoolSettings = collections.namedtuple(
         'autoscale', 'node_fill_type', 'remote_access_control',
         'certificates', 'prometheus', 'upload_diagnostics_logs_on_unusable',
         'container_runtimes_install', 'container_runtimes_default',
-        'per_job_auto_scratch',
+        'per_job_auto_scratch', 'environment_variables',
+        'environment_variables_keyvault_secret_id',
     ]
 )
 SSHSettings = collections.namedtuple(
@@ -1328,6 +1329,10 @@ def pool_settings(config):
         container_runtimes_default=cr_default,
         per_job_auto_scratch=_kv_read(
             conf, 'per_job_auto_scratch', default=False),
+        environment_variables=_kv_read_checked(
+            conf, 'environment_variables', default={}),
+        environment_variables_keyvault_secret_id=_kv_read_checked(
+            conf, 'environment_variables_keyvault_secret_id'),
     )